rename base to seq2seq base
