+++ dataquality/integrations/seq2seq/hf.py
@@ -3,7 +3,7 @@ from warnings import warn
 
 from transformers import GenerationConfig, PreTrainedModel, PreTrainedTokenizerFast
 
-from dataquality.loggers.logger_config.seq2seq.base import seq2seq_logger_config
+from dataquality.loggers.logger_config.seq2seq.seq2seq_base import seq2seq_logger_config
 from dataquality.schemas.split import Split
 from dataquality.schemas.task_type import TaskType
 from dataquality.utils.helpers import check_noop
