+++ dataquality/loggers/data_logger/seq2seq/chat.py
@@ -1,4 +1,4 @@
-from dataquality.loggers.data_logger.seq2seq.base import Seq2SeqDataLogger
+from dataquality.loggers.data_logger.seq2seq.seq2seq_base import Seq2SeqDataLogger
 from dataquality.loggers.logger_config.seq2seq.chat import (
     Seq2SeqChatLoggerConfig,
     seq2seq_chat_logger_config,
