+++ dataquality/loggers/data_logger/seq2seq/completion.py
@@ -1,4 +1,4 @@
-from dataquality.loggers.data_logger.seq2seq.base import Seq2SeqDataLogger
+from dataquality.loggers.data_logger.seq2seq.seq2seq_base import Seq2SeqDataLogger
 from dataquality.loggers.logger_config.seq2seq.completion import (
     Seq2SeqCompletionLoggerConfig,
     seq2seq_completion_logger_config,
