+++ dataquality/loggers/model_logger/seq2seq/seq2seq_base.py
@@ -4,7 +4,7 @@ import numpy as np
 import pyarrow as pa
 from scipy.special import log_softmax
 
-from dataquality.loggers.logger_config.seq2seq.base import (
+from dataquality.loggers.logger_config.seq2seq.seq2seq_base import (
     Seq2SeqLoggerConfig,
     seq2seq_logger_config,
 )
